pyproject.toml
[project]
name = "sg-draw"
version = "0.1.0"
description = "Photo to abstract art via scene graph"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.2.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.1",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "scipy>=1.9.0",
    "timm>=0.9.0",
    "opencv-python>=4.0.0",
]
